[DM/FIXUP] add cortex-m23's cpuport.h
