+++ libcpu/arm/cortex-m23/cpuport.c
@@ -14,34 +14,7 @@
 
 #include <rtthread.h>
 
-struct exception_stack_frame
-{
-    rt_uint32_t r0;
-    rt_uint32_t r1;
-    rt_uint32_t r2;
-    rt_uint32_t r3;
-    rt_uint32_t r12;
-    rt_uint32_t lr;
-    rt_uint32_t pc;
-    rt_uint32_t psr;
-};
-
-struct stack_frame
-{
-    /* r4 ~ r7 low register */
-    rt_uint32_t r4;
-    rt_uint32_t r5;
-    rt_uint32_t r6;
-    rt_uint32_t r7;
-
-    /* r8 ~ r11 high register */
-    rt_uint32_t r8;
-    rt_uint32_t r9;
-    rt_uint32_t r10;
-    rt_uint32_t r11;
-
-    struct exception_stack_frame exception_stack_frame;
-};
+#include "cpuport.h"
 
 /* flag in interrupt handling */
 rt_uint32_t rt_interrupt_from_thread, rt_interrupt_to_thread;
